Fix/readme rpm (#69)
* feat: rpm package repo

* feat: rpm package repo

* Develop (#68)

* Feature/rpm repo (#67)

* feat: rpm package repo

* feat: rpm package repo

* release: develop


* fix: added rpm install info

* feat: detect config by smbios info

* fix: sqlite static linking

* fix: config name to match device model

* log levels

* fix: service log level

* fix: service files

+++ res/uninstall/linux.sh
@@ -4,13 +4,26 @@ set -e
 set -x
 
 echo "nsense Linux uninstaller script"
-systemctl stop nsense
-systemctl disable nsense
-systemctl disable nsense-sleep
-rm -f /etc/systemd/system/nsense*
 
 if [ -d /opt/nsense ]; then
     rm -rf /opt/nsense
+fi
+
+executable=$(readlink /proc/1/exe)
+bin=${executable##*/}
+
+if [ "$bin" = "systemd" ]; then
+    if systemctl status nsense 2>&1 1>/dev/null; then
+        systemctl stop nsense
+    fi
+    systemctl disable nsense
+    systemctl disable nsense-sleep
+
+    rm -f /etc/systemd/system/nsense*
+else
+
+    echo "ERROR: your init system seems isn't supported" 1>&2
+    exit 1
 fi
 
 systemctl daemon-reload
